Add some user services

+++ server/services/user_service.py
@@ -1,4 +1,5 @@
 from bson import ObjectId
+from ..models.login_request import LoginRequest
 from ..models.user_model import UserModel
 from ..models.signup_request import SignupRequest
 from ..extensions import mongo
@@ -13,7 +14,13 @@ def findAll():
 
 
 def findOne(user_id: str):
+    if not ObjectId.is_valid(user_id): return None
     filter = {"_id": ObjectId(user_id)}
+    return users_db.find_one(filter)
+
+
+def findByEmail(user_email: str):
+    filter = {"email": user_email}
     return users_db.find_one(filter)
 
 
@@ -31,7 +38,18 @@ def createUser(signup: SignupRequest):
     return user
 
 
+def validateUser(login: LoginRequest):
+    user = findByEmail(login.getEmail())
+    if not user: return None
+
+    hashedPwd = hashPwdAndSalt(login.getPassword(), "asdf")
+    if user["password"] == hashedPwd:
+        return user
+    return None
+
+
 def delete(user_id: str):
+    if not ObjectId.is_valid(user_id): return None
     filter = {"_id": ObjectId(user_id)}
     result = users_db.delete_one(filter)
     return result.deleted_count
